pyproject.toml
[project]
name = "fh-swifty-chatbot"
version = "0.10.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "chainlit>=2.7.2",
    "fastmcp>=2.13.0.2",
    "ipykernel>=6.30.1",
    "langchain-community>=0.3.29",
    "langchain-mcp-adapters>=0.1.12",
    "langchain-openai>=0.3.33",
    "langchain-tavily>=0.2.11",
    "langgraph>=0.6.7",
    "neo4j>=6.1.0",
]

[tool.setuptools.packages.find]
include = ["fh-swifty-chatbot*"]

[tool.uv.workspace]
members = [
    "crawler",
]

[dependency-groups]
notebook = [
    "crawl4ai>=0.7.4",
    "ipykernel>=6.30.1",
    "playwright>=1.55.0",
    "selenium>=4.35.0",
    "unstructured>=0.18.14",
]
